refactor(wallet): move WalletLock component to a new directory and update imports
